refactor: (#327) 상세 게시글을 패치하는 컴포넌트 분리 및 서스펜스 적용 상세 게시글 MOCK 데이터 잘못 되어있던 것 수정, 사용하지 않는 코드 삭제, MOCK 데이터 파일 postList와 post를 병합
