+++ frontend/src/mocks/postList.ts
@@ -7,7 +7,7 @@ import {
   rest,
 } from 'msw';
 
-import { MOCK_POST_LIST } from '@mocks/mockData/postList';
+import { MOCK_POST_LIST } from '@mocks/mockData/post';
 
 export const mockPostList = [
   rest.get('/posts', (req, res, ctx) => {
