+++ frontend/src/pages/ErrorBoundary.tsx
@@ -2,34 +2,31 @@ import { Component, ErrorInfo, ReactNode } from 'react';
 
 interface ErrorBoundaryProps {
   children: ReactNode;
-  fallback: ReactNode;
 }
 
 interface ErrorBoundaryState {
   hasError: boolean;
+  errorMessage: string;
 }
 
 class ErrorBoundary extends Component<ErrorBoundaryProps, ErrorBoundaryState> {
-  #errorMessage = '';
-
   constructor(props: ErrorBoundaryProps) {
     super(props);
-    this.state = { hasError: false };
+    this.state = { hasError: false, errorMessage: '' };
   }
 
   static getDerivedStateFromError(error: Error) {
-    return { hasError: true };
+    return { hasError: true, errorMessage: error.message };
   }
 
   componentDidCatch(error: Error, errorInfo: ErrorInfo) {
     // You can also log the error to an error reporting service
     window.console.log(error, errorInfo);
-    this.#errorMessage = error.message;
   }
 
   render() {
     if (this.state.hasError) {
-      return <div>{this.#errorMessage}</div>;
+      return <div>{this.state.errorMessage}</div>;
     }
 
     return this.props.children;
