+++ frontend/src/pages/post/PostDetail/PostDetailFetcher/style.ts
